requirements.txt
argon2-cffi==25.1.0
django-celery-beat==2.8.1
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
dj-database-url==3.0.1
fido2==2.0.0
importlib-metadata==8.0.0
jaraco-functools==4.3.0
jaraco.collections==5.1.0
django-phonenumber-field[phonenumbers]
pip-chill==1.0.3
platformdirs==4.2.2
psycopg2-binary==2.9.10
python-decouple==3.8
python-slugify==8.0.4
redis==6.4.0
requests
tomli==2.0.1
whitenoise==6.10.0
